Adjust storage and processing in SystemPacketBuffer  and associated code for large payloads.
Extend uint16_t type variables to size_t for the APIs
and all applicable places.

+++ src/app/BufferedReadCallback.cpp
@@ -65,7 +65,7 @@ CHIP_ERROR BufferedReadCallback::GenerateListTLV(TLV::ScopedBufferTLVReader & aR
     //
     // To avoid that, a single contiguous buffer is the best likely approach for now.
     //
-    uint32_t totalBufSize = 0;
+    size_t totalBufSize = 0;
     for (const auto & packetBuffer : mBufferedList)
     {
         totalBufSize += packetBuffer->TotalLength();
